app/requirements.txt
python-dotenv
pinecone-client
openai
tiktoken
streamlit
